Core: fix compilation issue after the merge
