+++ src/main/java/org/semux/Kernel.java
@@ -11,8 +11,6 @@ import java.io.IOException;
 import java.net.InetAddress;
 import java.util.Arrays;
 import java.util.Map;
-import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 import javax.xml.parsers.ParserConfigurationException;
@@ -30,7 +28,6 @@ import org.semux.core.BlockchainFactory;
 import org.semux.core.Genesis;
 import org.semux.core.PendingManager;
 import org.semux.core.SyncManager;
-import org.semux.core.Transaction;
 import org.semux.core.Wallet;
 import org.semux.crypto.Hex;
 import org.semux.crypto.Key;
@@ -45,7 +42,6 @@ import org.semux.net.ChannelManager;
 import org.semux.net.NodeManager;
 import org.semux.net.PeerClient;
 import org.semux.net.PeerServer;
-import org.semux.util.ByteArray;
 import org.semux.util.Bytes;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
