+++ src/main/java/org/semux/core/BlockchainImplV2.java
@@ -26,6 +26,8 @@ import org.apache.commons.lang3.tuple.Pair;
 import org.ethereum.vm.client.BlockStore;
 import org.semux.config.Config;
 import org.semux.config.Constants;
+import org.semux.core.codec.BlockCodec;
+import org.semux.core.codec.BlockCodecV2;
 import org.semux.core.exception.BlockchainException;
 import org.semux.core.state.AccountState;
 import org.semux.core.state.AccountStateImplV2;
@@ -40,8 +42,6 @@ import org.semux.db.BatchManager;
 import org.semux.db.BatchName;
 import org.semux.db.BatchOperation;
 import org.semux.db.Database;
-import org.semux.db.DatabaseFactory;
-import org.semux.db.DatabaseName;
 import org.semux.db.DatabasePrefixesV2;
 import org.semux.db.DatabaseVersion;
 import org.semux.util.ByteArray;
