--- src/main/java/org/semux/core/TransactionDecoder.java
@@ -1,8 +1,0 @@
-package org.semux.core;
-
-public interface TransactionDecoder {
-
-    Transaction decode(byte[] bytes);
-
-    Transaction decodeUnsigned(byte[] bytes);
-}
