--- src/main/java/org/semux/core/TransactionEncoder.java
@@ -1,9 +1,0 @@
-package org.semux.core;
-
-public interface TransactionEncoder {
-
-    byte[] encode(Transaction transaction);
-
-    byte[] encodeUnsigned(Transaction transaction);
-
-}
