+++ src/main/java/org/semux/core/codec/BlockCodecV1.java
@@ -4,7 +4,7 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.core;
+package org.semux.core.codec;
 
 public class BlockCodecV1 implements BlockCodec {
