+++ src/main/java/org/semux/core/codec/BlockDecoder.java
@@ -4,11 +4,12 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.core;
+package org.semux.core.codec;
 
 import java.util.List;
 
 import org.apache.commons.lang3.tuple.Pair;
+import org.semux.core.Block;
 import org.semux.crypto.Key;
 
 public interface BlockDecoder {
