+++ src/main/java/org/semux/core/codec/BlockEncoderV1.java
@@ -4,12 +4,15 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.core;
+package org.semux.core.codec;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.commons.lang3.tuple.Pair;
+import org.semux.core.Block;
+import org.semux.core.Transaction;
+import org.semux.core.TransactionResult;
 import org.semux.crypto.Key;
 import org.semux.util.SimpleEncoder;
