+++ src/main/java/org/semux/core/codec/TransactionDecoder.java
@@ -4,11 +4,13 @@
  * Distributed under the MIT software license, see the accompanying file
  * LICENSE or https://opensource.org/licenses/mit-license.php
  */
-package org.semux.core;
+package org.semux.core.codec;
 
-public interface BlockCodec {
+import org.semux.core.Transaction;
 
-    BlockEncoder encoder();
+public interface TransactionDecoder {
 
-    BlockDecoder decoder();
+    Transaction decode(byte[] bytes);
+
+    Transaction decodeUnsigned(byte[] bytes);
 }
