+++ src/main/java/org/semux/core/codec/TransactionDecoderV1.java
@@ -1,6 +1,15 @@
-package org.semux.core;
+/**
+ * Copyright (c) 2017-2018 The Semux Developers
+ *
+ * Distributed under the MIT software license, see the accompanying file
+ * LICENSE or https://opensource.org/licenses/mit-license.php
+ */
+package org.semux.core.codec;
 
 import org.semux.Network;
+import org.semux.core.Amount;
+import org.semux.core.Transaction;
+import org.semux.core.TransactionType;
 import org.semux.util.SimpleDecoder;
 
 public class TransactionDecoderV1 implements TransactionDecoder {
