+++ src/main/java/org/semux/core/codec/TransactionEncoderV1.java
@@ -1,5 +1,13 @@
-package org.semux.core;
+/**
+ * Copyright (c) 2017-2018 The Semux Developers
+ *
+ * Distributed under the MIT software license, see the accompanying file
+ * LICENSE or https://opensource.org/licenses/mit-license.php
+ */
+package org.semux.core.codec;
 
+import org.semux.core.Transaction;
+import org.semux.core.TransactionType;
 import org.semux.util.SimpleEncoder;
 
 public class TransactionEncoderV1 implements TransactionEncoder {
