+++ src/main/java/org/semux/core/state/AccountEncoderV1.java
@@ -1,3 +1,9 @@
+/**
+ * Copyright (c) 2017-2018 The Semux Developers
+ *
+ * Distributed under the MIT software license, see the accompanying file
+ * LICENSE or https://opensource.org/licenses/mit-license.php
+ */
 package org.semux.core.state;
 
 import org.semux.util.SimpleEncoder;
