+++ src/test/java/org/semux/bench/BlockchainPerformance.java
@@ -18,8 +18,6 @@ import org.semux.config.Constants;
 import org.semux.config.DevnetConfig;
 import org.semux.core.Amount;
 import org.semux.core.Block;
-import org.semux.core.BlockEncoder;
-import org.semux.core.BlockEncoderV1;
 import org.semux.core.BlockHeader;
 import org.semux.core.Blockchain;
 import org.semux.core.BlockchainFactory;
@@ -27,6 +25,8 @@ import org.semux.core.Genesis;
 import org.semux.core.Transaction;
 import org.semux.core.TransactionResult;
 import org.semux.core.TransactionType;
+import org.semux.core.codec.BlockEncoder;
+import org.semux.core.codec.BlockEncoderV1;
 import org.semux.crypto.Key;
 import org.semux.crypto.Key.Signature;
 import org.semux.rules.TemporaryDatabaseRule;
