+++ src/test/java/org/semux/bench/CompressPerformance.java
@@ -15,7 +15,6 @@ import org.semux.config.MainnetConfig;
 import org.semux.core.Block;
 import org.semux.core.Blockchain;
 import org.semux.core.BlockchainFactory;
-import org.semux.core.BlockchainImpl;
 import org.semux.core.Genesis;
 import org.semux.db.LeveldbDatabase.LeveldbFactory;
 import org.semux.net.msg.consensus.BlockMessage;
