+++ src/test/java/org/semux/consensus/SemuxSyncTest.java
@@ -44,7 +44,6 @@ import org.semux.core.Amount;
 import org.semux.core.Block;
 import org.semux.core.Blockchain;
 import org.semux.core.BlockchainFactory;
-import org.semux.core.BlockchainImpl;
 import org.semux.core.BlockchainImplV2;
 import org.semux.core.Fork;
 import org.semux.core.Genesis;
