+++ src/test/java/org/semux/core/BlockTest.java
@@ -21,6 +21,9 @@ import org.semux.Network;
 import org.semux.config.Config;
 import org.semux.config.Constants;
 import org.semux.config.DevnetConfig;
+import org.semux.core.codec.BlockDecoderV1;
+import org.semux.core.codec.BlockEncoder;
+import org.semux.core.codec.BlockEncoderV1;
 import org.semux.crypto.Key;
 import org.semux.crypto.Key.Signature;
 import org.semux.util.Bytes;
