+++ src/test/java/org/semux/core/BlockchainImplMigrationTest.java
@@ -24,8 +24,8 @@ import org.semux.Network;
 import org.semux.config.Config;
 import org.semux.config.Constants;
 import org.semux.config.TestnetConfig;
-import org.semux.core.bip32.extern.Hex;
 import org.semux.core.state.Delegate;
+import org.semux.crypto.Hex;
 import org.semux.db.DatabaseFactory;
 import org.semux.db.DatabaseVersion;
 import org.semux.db.LeveldbDatabase;
