+++ src/test/java/org/semux/core/BlockchainImplTest.java
@@ -28,6 +28,9 @@ import org.semux.config.AbstractConfig;
 import org.semux.config.Config;
 import org.semux.config.Constants;
 import org.semux.config.DevnetConfig;
+import org.semux.core.codec.BlockDecoderV1;
+import org.semux.core.codec.BlockEncoder;
+import org.semux.core.codec.BlockEncoderV1;
 import org.semux.crypto.Key;
 import org.semux.rules.TemporaryDatabaseRule;
 import org.semux.util.Bytes;
