+++ src/test/java/org/semux/gui/SemuxGuiTest.java
@@ -22,12 +22,9 @@ import org.mockito.Mockito;
 import org.semux.KernelMock;
 import org.semux.Network;
 import org.semux.config.Config;
-import org.semux.config.Constants;
-import org.semux.config.DevnetConfig;
 import org.semux.consensus.SemuxSync;
 import org.semux.core.Blockchain;
 import org.semux.core.BlockchainFactory;
-import org.semux.core.BlockchainImpl;
 import org.semux.core.Genesis;
 import org.semux.core.Wallet;
 import org.semux.db.LeveldbDatabase;
